BM-181: Add a `main-status-check` CI job and expand codeowners (#25)
This PR adds a `main-status-check` CI job, which we can use as the job
that's required to pass before a PR can merge. It also examples the
codeowners to allow more people to be reviewers on PRs.

There doesn't appear to be a team that only contains devs working on
Boundless, so I just added engineers by username for now. We should
create that team. I included @zeroecco so that @mothran and him can have
enough privileges to merge PRs to Bento without a third person.