BM-266: Resubmitting merkle roots cleanup (#83)
This PR allows for cheap resubmission of timed-out submit_merkle
batches. It does so by checking if the root is already present and
skipping to fulfill_batch. Most useful for manual triggering of
re-submission of batches.

It also adds docs on how to go about manual re-submission of orders in
the Broker docs.