BM-272: Fix expired status (#100)
A proof request is expired when the current block number is greater than
the sum `startBidding` + `timeout` and it's not fulfilled. A proof
request expiration is stored into the contract state only during a
`lockinRequest`. As such, querying the status of an expired proof
request yields an `expired` response only if the proof request has been
locked in, and `unknown` otherwise. To minimise this limitation, the
`get_status` method accept an optional `expires_at` parameter the caller
should provide. The same parameter is made mandatory for the
`wait_for_proof_fulfillment` method.

---------

Co-authored-by: Victor Graf <victor@risczero.com>