BM-308: Refactor fulfillment and slashing paths (#135)
Refactor the fulfillment and slashing methods to allow for a proof to be
delivered, and a request to be marked as fulfilled, without sending
payment if the proof is valid but the requirements for payment are not
met (e.g. it is not locked to the prover that sent the fulfill
transaction). This behavior is controlled by setting the
`requirePayment` flag on the `Fulfillment` struct.

When `requirePayment` is set to `true`, the behavior of the fulfillment
methods is as before. When set to `false`, the method will still revert
on errors in validating the proofs, but it will emit a
`PaymentRequirementsFailed` event with the encoded error that prevented
payment from being issued. In some cases, the prover can still collect
payment later by submitting a new fulillment transaction with correct
parameters.

This PR also refactors the tests and adds a number of tests to attempt
to cover all the different paths that now exist in the contract. More
work could be done here, but it is at a point that I think we should
call it for this PR.

This PR also fixes a few minor issues:
* Sending value to the `submitRequest` method did not emit a `Deposit`
event.
* Comments on methods were made as "//" comments instead of "///"
docstring comments.
* Adds a "slashing burn fraction" constant that we can adjust to send
part of the burn amount to the client.
* Adds an `InvalidSignature` error to use in place of a string.
* Change the way `ProofMarketService::submit_request` handles deposits.
* Use trial-decoding to attempt to decode Solidity errors with the
errors from multiple interfaces.