BM-5: Remove uint192 and uint96 from public interfaces (#160)
This PR removes the usages from `uin192` and `uint96` from public
interfaces. Using them was a premature optimization as we do not store
these types in EVM state and it is unclear / doubtful that using them
provides optimization advantages otherwise. It is clear that it adds a
bit of friction to the UX.

This PR also removes the last usages of strings in EVM `require`
statements, and refactors usage of `anyhow` in the request validation
function.

Bonus: Expanding the request ID to 256 bits gives us 64 bits (currently
enforced to be zero) to use as needed in the future. In particular, this
gives us a potential place to insert versioning flags if needed.

Resolves #92
Resolves BM-280
Resolves BM-5