BM-122: Use request digest to ensure the locked/priced request matches the one being filled (#163)
An issue with the approach on `main` is: if the client signs a request
with a given ID, it expires with no bids, and they sign a new request
with the same ID and different requirements, which gets a lockin, the
prover is not actually bound to fulfill the one that was locked in.

This PR addresses this issue by using the EIP-712 digest of the request
in the request lock, in the transient pricing mechanism, and in the
assessor journal.

Note that this PR brings-to-light what is more-or-less an existing issue
with the implementation. Because more than one request can have the same
ID, is is not a unique identifier of all the requests a client may have
signed. In some places, we use the request ID when we should really be
using the request digest. This is in tension with the dev/user
experience we'd like to provide, which is to have a single simple
identifier for their request. Request ID has a number of advantages as
this identifier, but also has this issue.