BM-358: Rename ProofMarket to BoundlessMarket (#168)
This PR renames the `ProofMarket` contract to `BoundlessMarket` and
propogates those changes to all the relevant places. Additionally,
`ProvingRequest` is renamed to `ProofRequest` and in prose the "proof
request" is made simply to be "request" where is seems clear from the
context.

This PR is targeted to merged on top of
https://github.com/boundless-xyz/boundless/pull/163 to avoid merge
conflicts, as both PRs touch quite a few files.