BM-288: Use safe upgrade (#151)
This PR refactor the `manage` script to deploy and upgrade the
ProofMarket using the (safe) Upgrades. ~~It adds a new file
`ProofMarketLastDeployment.sol` that matches with the last deployed
contract and it is used as a reference contract for both upgrades and
test upgrades. Note that this file should always match with the last
deployment.~~

Closes #109 
Closes BM-289