BM-385: Improvements and fixes for the boundless market CLI and the fulfill tool (#192)
As I was going through the docs / testing things out, I put together
some fixes and improvements to the CLI tools. The biggest change is to
run preflight by default before submit request, rather than as a
separate dry-run. I added functionality to the execute command that does
part of the what dry-run notion was covering before, and is generally
useful to validate a request YAML file (I used it to debug an issue I
had in updating it)

---------

Co-authored-by: Angelo Capossele <angelocapossele@gmail.com>