BM-396: Return `expires_at` as part of submitting a request (#199)
The `boundless-market` method `wait_for_request_fulfillment` requires a
`u64` representing the expiration of the request (expressed as the block
number). The current code uses `request.expires_at()`. The problem is
that typically the request.bidding_start is modified (if initialized to
0) by the client during request submission, as it has access to the
current block number.
This PR returns the correct `expires_at` value, so that can be used in
`wait_for_request_fulfillment`