BM-409: Use random request index by default (#210)
I realized in some final testing that using the wallet nonce is no
longer a great option with the fact that a user may use a combination of
on and off-chain requests. Using sequential IDs is more gas-efficient,
but I think for now we can set the default to random until we can
revisit this.