BM-420: Use get_receipt in `price_and_fulfill_batch` and `submit_merkle_and_fulfill` (#221)
After some time, the broker starts to fail when calling
`submit_merkle_and_fulfill`. However, the tx gets through, so it is
somewhere in alloy that triggers an error. One theory is that the
`watch` method awaits immediately on the `PendingTransaction` while the
`get_receipt` instead waits until a tx receipt or an error. It also
seems that only the `submit_merkle_and_fulfill` is triggering this error
and instead the `lockin` is not (using already `get_receipt`).

Co-authored-by: Parker Thompson <parker@risczero.com>