[BM-456] Refactor logic to types and libraries, add tests (part 1) (#253)
Splitting this into two parts as the PR is getting big.

List of changes:
1/ lockin => lockRequest
2/ Introduce RequestId alias for uint256
3/ Introduce types folder (for types and their libraries e.g. Offer,
ProofRequest) and libraries folder (for libraries e.g. MerkleProofish)
4/ Basic tests for each type
5/ Remove duplicated comments from BoundlessMarket.sol

Left for part 2:
1/ Moving all ProofRequest hashing/signing logic to the ProofRequest
library. I need to spend some time understanding `_hashTypedDataV4` and
how it works with contract upgradability, as I can not call it directly
from the library.
2/ Fulfilment and slashing. I'm working on two other tasks (BM-443 and
BM-450) that will touch the logic here.
3/ Fixing the upgradability tests - haven't got to this yet

---------

Co-authored-by: Victor Snyder-Graf <victor@risczero.com>