BM-475: Refactor aggreagtion logic using the new aggregation guest (#264)
Uses the new incremental tree-building logic to refactor and improve
aggregation. One of the key goals of this PR is to reduce the number of
proving round trips required for aggregation. Instead of a singleton and
join for each proof in a batch, there is now a single run of the
set-builder when we want to immediately finalize, and one execution per
round when we want to build incrementally.

---------

Co-authored-by: Angelo Capossele <angelocapossele@gmail.com>