BM-516: DB fuzz test (#297)
Unsuccessful in triggering the "database is locked" error, but have
noted the following from these tests:

- Block number is cast from a u64 to i64, which can cause invalid state
in the DB. Unlikely for any chain it will exceed `u32::MAX` block
numbers, but maybe safer to add the sanity check or handle this
differently
- `add_order` can hit a collision of order IDs. Perhaps checked in logic
of broker and how it's synced with chain data, but DB gets in a borked
state if collision, so we should be sure about any potential
conditions/DOS vector
- `get_aggregated_proofs` requires that a few optional fields on the
order are set, seems like they should be mandatory on the order, or
optional on the aggregate struct?
- Curious why the table is updated in this method, perhaps name should
be changed to make it clear there is a side effect, at least?
- On large enough input through the prop test, seems to consistently hit
`called `Result::unwrap()` on an `Err` value: SqlErr(PoolTimedOut)` on
any operation.


Maybe it's the case that some condition of the broker running causes the
db is locked error, like a starved executor or sth like that, but #266
didn't catch it either. Also very possible I'm not testing accurate
flows or missing code paths. I tried to match existing patterns, but
hard to match exactly.

(draft until we resolve what we want to do with TODOs)