BM-535: Disable docker build for order-generator-zeth, fix retry bug (#313)
By default building zeth requires docker in order to create a
reproducible build of the guests. This means that when building
order-generator-zeth in docker, the build process then tries to spawn
Docker within Docker when it compiles zeth. Docker within Docker [is not
recommended](https://stackoverflow.com/a/33003273), and requires Docker
to be run in privileged mode which causes problems for our deployments.

Also fixing a bug where the Zeth build would fail, reducing the number
of retries remaining, but the call to getBlock would succeed which
resets the retry count to 0, causing the task to loop forever.