BM-570: bump zkvm version for backported alignment fix (#341)
Just changes locked versions.

This also unfortunately has to copy the serialization/deserialization
code from https://github.com/risc0/risc0/pull/2834 and do some clones on
each end (which are both necessary). Unfortunately redis for some reason
doesn't have a `[T; N]` impl, and only a `&[T; N]`, which doesn't work
because the lifetime needs to be static. This can likely be removed with
a newtype, but changing the type might be a breaking change, so avoiding
making that change in this PR.