BM-459: Explore options for delivering unaggregated proofs (#306)
**Part 1**

This first pass only covers the contracts side of things. The idea is to
add an optional field `Selector` to the `Requirements` of a proof
request. The `selector` is also committed by the assessor in the
`AssessorJournal` (it only commits the ones actually required as
otherwise any selector would suffice). **Currently the broker skips any
request that specifies a selector.**

**Part 2** should instead focus on the prover side: the broker must
check that it can support the required selector (if set) and generate
the seal accordingly

---------

Co-authored-by: Victor Snyder-Graf <victor@risczero.com>