BM-544: Support smart contract signatures (#346)
Adds smart contract signature support

One funny thing, due to the way our "Client" object was created in our
tests we were essentially sending requests from Smart Contracts already.
This also meant that all the tests started failing once we had real SC
support via EIP-1271. As a result I had to refactor Client, and took the
opportunity to split things into different files while I was doing this.
So lots of changes to tests.

One other note, to make this work for lockRequestWithSignature, it
required adding a new field "proverAddress" to the interface. Without
this it is not possible to support smart contract signatures for that
method.

---------

Co-authored-by: Victor Graf <victor@risczero.com>