BM-510: Stake rollover between requests with same id (#357)
If two requests have the same id, and one is locked and not fulfilled.
The fulfiller of the second request is eligible for a portion of the
slashed stake. Note a request can only ever be locked once- without this
property this would be a lot more complex to implement.

Some edge cases here:
1/ If the first request fully expires, then someone calls slash before
the second request is fulfilled, then the second request fulfiller will
not receive a portion of slashed stake
2/ If the second request is cheaper than the first request, the client
may be eligible a refund (they were charged a high amount at lock time,
then the fulfillment was ultimately cheaper)

---------

Co-authored-by: Victor Graf <victor@risczero.com>