BM-416: Use timestamps instead of block numbers for request information (#375)
Change the auction timing (e.g. `offer.biddingStart` and
`offer.timeout`) to use timestamps instead of block numbers. This should
simplify the experience for developers and make it more consistent
across different blockchains.

As a drive-by fix, this fixes the logic for determining how much gas and
stake is assigned to pending orders.