BM-606: Set price to zero after lock deadline (#378)
In recent work, we split the request timeout into two parts: a lock
timeout and a timeout. In particular, the lock timeout is the moment
where any prover that has locked the request is considered to have
failed their obligation, loses exclusivity, and is slashable.

On `main` right now, the price of a request continues to increase during
the locked period, and so if the lock expires it is likely that the
price has increased to `maxPrice` (and more generally, it is no less
than the original lockin price). Because of exclusivity, if a prover
submits their bid, there is no opportunity for other provers to indicate
their bid until the lock has expired, and the price may have risen past
where they would have bid. This impacts price discovery, and can cause
the client to end up paying more than the "true" market price defined as
the minimum price that a prover capable of filling the request would
accept. Note that this "true" price is not directly observable.

Adding to this, since (unburned portion of) the stake is added to the
reward to any prover that fulfills after the lock deadline, in the case
that the locked prover fails, there is some notion that if the stake is
high enough, the stake alone is sufficient to motivate a prover to fill
the request. It is likely that this is the case for a large class of
requests.

After much discussion, we've decided to address the issue above by
setting the price paid by the client for a request delivered after the
lock deadline and before the final request deadline to zero. As a
reward, the prover will still obtain any unburned stake taken from a
prover that failed the request, if another prover previously locked it.

As with most market mechanisms, it is impossible to tell whether this is
the correct choice until we have data from a live market with funds
flowing through it. When we have that data, we may revisit this.

---------

Co-authored-by: pote.eth <81638931+willpote@users.noreply.github.com>
Co-authored-by: Pote <pote@risczero.com>