BM-629: Add retry policy to Eth RPC client in order stream (#401)
We've been seeing periodic failures on RPC requests made by the client
in order stream. In particular, the balance check is failing sometimes
which results in a 500 error being sent to the client. In order to
handle this more gracefully more often, this PR adds a retry layer to
the RPC client used by order-stream.