BM-688: use FOUNDRY_OUT env var (#445)
For some reason the `out` field of the `profile.reference-contract` in
the `foundry.toml` does not get used. This PR forces its usage with the
env var `FOUNDRY_OUT`
It also fixes a crate version mismatch, `risc0-groth16` by pinning and
forcing to use a specific version. This can be removed once upgrading
the zkvm to v2.0