BM-644: Setup for aws deployments (#448)
Initial setup for deploying from the Boundless repo. Goals here are to:
1. allow us to update our prover and other infra with just 1 PR,
speeding development
2. house business logic and the infra code in one repo, simplifying
development of AWS apps, where business logic and infra are often
tightly coupled
3. move us closer to continuous deployments

Note there are three pulumi apps in this initial PR:
1. The `bootstrap` app deploys a role to each service account. This role
is used by the `pipelines` app to deploy to each service account.
4. The `pipelines` app contains the Pulumi state bucket and keys, and
AWS CodePipeline deployment pipelines that deploy to staging/prod when
code is merged. Currently it contains one pipeline that deploys a
`sample` app.
5. The `sample` app is intended as a sample/template app, showing local
develop, and secret management.

The AWS account structure is that we have an ops account that contains
the deployment pipelines, Pulumi keys, and Pulumi state bucket. This
account deploys cross-account to the staging account and the prod
account. We also have a development account, that each developer can use
to deploy their own stacks for testing.