BM-693: use cpus instead of cpu_count (#451)
ubuntu and linux use "cpus". "cpu_count" is for windows only.

--cpu-count and --cpu-percent is for the HyperV isolation on Windows. When you run a Windows container you have two options.

Using process isolation: You must have the same Windows version in the container than on the host
Using HyperV isolation: You will have a virtualized environment, so the container is actually a VM which requires some CPUs. I don’t remember the default number of CPUs, but if it is 1 or 2, but you have 12 cores, you may want to use more, so you can set --cpu-count.