BM-684: rename batch_size to min_batch_size (#443)
This was confusing to me, but also was an assumption of an external
party that this would limit the amount of orders locked. Indicating this
is a minimum might make it more clear what it's configuring.

This is semver breaking, but backwards compatible with the previous
`broker.toml` naming, such as to not break previous configs on update.
Can also have this be non-api breaking, but probably better to break now
rather than later.