BM-705: Assorted audit fixes (#465)
#459

Note there are two recommendations not implemented:

1. **In depositStakeWithPermit(), the call to the ERC-20 permit()
function is wrapped in a try-catch block. This could lead to confusing
error reporting when the transaction later reverts from calling
safeTransferFrom().**

try/catch for permit, we are following the recommendations from OZ here:
https://docs.openzeppelin.com/contracts/4.x/api/token/erc20#security_considerations

2. **withdrawFromTreasury() can be simplified with a call to
_withdraw().**

For withdrawFromTreasury, we withdraw from address(this) but send to
msg.sender, so _withdraw can not be used.