BM-707: Use read and write fram I/O functions in assessor (#461)
Currently on `main`, the assessor input is written/read using a custom
length-value encoding. The `write_frame` and `read_frame` functions now
exist to handle this parttern. This PR refactors to drop the manual
version of this.

It also renames `AssessorInput::to_vec` to `encode` and adds a `decode`
function to contain the usage of postcard to an opaque API, in case we
need to change it later.