BM-676: refactor order picker capacity logic (#433)
Current issue is that the order picker doesn't really have a mechanism
to update this reserved capacity in the case of a failed lock operation.
This refactors the logic such that if idle and no pricing tasks
scheduled, this reserved capacity/time is reset to current time.

Also updates the logic of the pricing function to decrement the
reservation if the order fails the order pricing checks, as there was
this similar race condition there as well.