BM-40: Account for estimated gas costs when pricing orders (#486)
On `main` , gas costs are not taken into account when pricing an order.
This is a significant issue as gas costs often dwarf the computational
costs of a proof when the proof is not sufficiently large.

This PR improves the basic gas estimation we currently use for gas
balance management, and uses it to add the gas costs into the pricing
function used for picking orders.

It also updates the `ChainMonitor` to refactor how it gets the block
timestamp, and adding a call to get the gas price estimate (as reported
by the `eth_gasPrice` RPC) along side the call to get the latest block
up to every X seconds.

Note this is a significant change to how pricing is done. We also need
to account for these changes on the client side when building a request.
This PR covers the broker changes, and I will create a follow-up PR with
the client changes to make it easier to set a reasonable price on a
given request.