BM-757: Use cert in prod (#502)
Now that the cert has been created and validated, we can enable the
order stream is prod to use it.

ACM_CERT_ARN is not necessary as in this design the cert is fully
managed by Pulumi, so we can get the ARN from the Pulumi resource.

This should be the last change to set up the DNS, though I'm unable to
test locally since this is prod only, so can't be sure.