BM-772: Resume slashing RISC Zero provers using the RISC Zero slasher (#518)
We operate a slashing bot on the Boundless testnet, and due to
difficulties with the freezing system that ultimately would cause
requestors to have a bad experience, we temporarily removed RISC Zero
provers from the list of slashing targets. We have not removed freezing,
and so now will slash our own provers again.