BM-783: Add an estimated upper bound on gas costs to order generator max price (#529)
At the moment, during Sepolia gas spikes the order generator is not
applying a high enough max price, since the price will not cover the gas
costs for the prover. This PR adds a first implementation of adding the
gas price to the offer from the generator.

Note that I adjust the ramp up default, however this is overridden in
our Pulumi deployment configs.