BM-511 Pick slashed unexpired orders (#469)
Adds the following to allow the broker to detect when orders have been
slashed and use different pricing logic

- New tasks in MarketMonitor that monitors for orders being locked and
filled by other provers. This allows the broker to get a better view of
the protocol state and it required for checking when an order has had
its lock expired but has not been filled

- Extends `update_orders_for_pricing` to also update orders that meet
the slashed+unexpired requirement

- Adds new config fields 
- - `fill_slashed_orders_altruistically` which skips profitability
checks for slashed orders if true
- - `mcycle_price_stake_token` which is the equivalent of `mcycle_price`
but expressed in stake tokens. This should be removed once the broker
has a view into the stake token price relative to the gas token

- Updates `price_order` logic to:
- - Correctly set the pricing estimate as a constant fraction of the
stake for slashed orders
- - Detect that the expiry is the order expiry and not the lock expiry
for slashed orders and the lockin stake is zero
- - Skip unprofitable slashed orders if configured to do so


Much of the churn in the diff is due to a refactor of the
boundless_market crate test utils. This has been moved to its own crate
allowing it to expose more code for testing using the market while also
being used within the market crate itself.

> If it is decided we don't want that refactor the branch up to commit
[1cb2e14](https://github.com/boundless-xyz/boundless/pull/469/commits/1cb2e140ba2252da5e9e3080e40e40084b3f9de4)
has all the order picking features prior to the refactor