BM-779: Refactor task retry and enable configurable policy (#526)
Closes BM-770
Closes BM-752

PR title and:
- Lowers the error of order stream failures to recoverable (including
the initial connection such that any recoverable restarts don't fault
the broker on restart if order stream down)
- Re-add dev mode defaulting to local prover removed in
https://github.com/boundless-xyz/boundless/pull/452
- Not sure if anything expects this to be removed, but localnet doesn't
work as is as it defaults to using bento

This implementation will warn and backoff retries for reconnecting.
Applies at least a small backoff to each retry, to avoid the spinning
retries that happen now, which more easily lead to rate limiting and
other issues. Tested manually that this handles no order stream on
startup, disconnect after startup, reconnections on localnet, and
reconnections to an order stream that returns an error on connect -- all
of which warn log on retries but do not fault the service.

---------

Co-authored-by: pote.eth <81638931+willpote@users.noreply.github.com>