BM-818: Adjust order generator parameters to reduce sETH churn in Zeth (#554)
We currently churn a lot of Sepolia ETH and the current order parameters
seem to place the max mcycle price much higher than it needs to be.

This PR adjusts the max price and min price per Mcycles. I also adjust
the staking amounts and the timeouts.

* Min price is set to 0 because it doesn't actually help to have it just
slightly higher than zero. In general, zero should be the default except
for cases where the sender has a reasonable idea of the correc price and
wants a faster or more precise auction.
* Max price is lowered to be 100x what we set as out mcycle price for
our prover. This is a reduction of 10x.
* Adjust ramp-up period to 10 minutes instead of 8m20s.
* Adjust stake on Zeth orders to 0.5 instead of 0.1. Intutively, this
job has a higher value and should have a higher stake than the simple
jobs.