BM-800: Fulfill after lock expire fixes. Change order table to support multiple fulfillment methods. (#542)
This PR firstly fixes some issues with the initial implementation for
fulfilling requests where the lock has expired. The first version had
two functional issues:

1. It attempted to fulfill the requests immediately when the
`RequestLocked` event was detected, rather than waiting until the lock
had expired. This has been resolved by storing a "target_timestamp"
value that represents when the lock expires, and only attempting to
fulfill after that point, similar to our logic for locking requests.

2. It attempted to call the contract to fulfill lock-expired requests
without pricing them. This is resolved by using `priceAndFulfillBatch`
when our batch contains lock expired orders.

These issues were missed due to us not having an E2E test. I've now
added an E2E test that covers the full workflow.

---

Secondly this PR makes some changes to the orders DB table aimed at
reducing complexity, race conditions, and setting us up for future
features like "fulfilling orders without locking" and "multiple requests
with same id".

The initial implementation monitored for `RequestLocked` events, and if
detected, updated the order's OrderStatus back to Pricing, to indicate
that we should reconsider the request for fulfilling after the lock
expired. This introduced a couple of issues:

1. Race Conditions: for example, if we updated the status back to
Pricing while the order was in the middle of proving, when the proving
workflow completed the status would be overwritten to
PendingAggregation, meaning the request is never re-priced.

2. Breaks the "assembly line" mental model, adding complexity. The
proving node originally processed orders like an assembly line;
market_monitor -> order_picker -> order_monitor -> proving ->
aggregation -> submission. Allowing an orders status to be set to a
previous status, breaks the simple pipeline model and complicates the
state transition graph, making things harder to debug and increasing
risks of bugs.

To address this, I have changed the orders table primary key to be of
type `{request_id}-{request_hash}-{fulfillment_type}`. Where fulfillment
type indicates whether we are intending to "Lock and Fulfill" the
request, or "Wait until the lock expires and fulfill" or "Fulfill
without locking the request". So now each request will result in
multiple rows in the orders table, where each row represents a proof
request +and+ a specific way to fulfill it, and each row goes through
the assembly line independently. (Note request hash is included in the
id to enable us to support multiple requests with the same id in the
future.)

---

Thirdly, this PR moves capacity tracking from the order-picker into the
order-monitor. Previously we would assign capacity during order-picking,
which in many cases in significantly before the time that we actually
need the capacity- as we must wait until the order is profitable enough
to lock, or wait until the lock expires to attempt to fulfill for
slashed stake. Now we only assign capacity when we hit the
`target_timestamp` and are about to either lock or prove the order
(depending on the fulfillment type)


---

A couple follow ups for the task will be:

Capacity limits for executor runs (this PR removes any limitations at
that stage)
Executor/proving caching - this design results in operations like
executor being repeated across each order row. We can cache the results
Adding a submitMerkleAndPriceAndFulfill method to the contract.
Currently fulfilling both locked and unlocked orders in one transaction
is not possible.